unwrap default + common folder
