+++ crates/infra/audit/src/lib.rs
@@ -36,9 +36,8 @@ pub use storage::{
 mod types;
 use tokio::sync::mpsc;
 use tracing::error;
-pub use types::{
-    BundleEvent, BundleId, DropReason, RejectedTransaction, Transaction, TransactionId,
-};
+pub use base_bundles::RejectedTransaction;
+pub use types::{BundleEvent, BundleId, DropReason, Transaction, TransactionId};
 
 /// Connects bundle event receivers to publishers.
 #[derive(Debug)]
